投稿作成機能を実装する

+++ db/migrate/20221231034139_create_laugh_logs.rb
@@ -2,7 +2,7 @@ class CreateLaughLogs < ActiveRecord::Migration[7.0]
   def change
     create_table :laugh_logs do |t|
       t.references :post, null: false, foreign_key: true
-      t.integer :button_pressed_time
+      t.float :button_pressed_time
 
       t.timestamps
     end
